Revert "feat(lib): add time.rs for monotonically increasig time (#244)"
This reverts commit f3055e5d735ced3ada4210f791682a4afe9b2dab.

+++ awkernel_lib/src/arch/std_common/delay.rs
@@ -44,7 +44,7 @@ impl Delay for super::StdCommon {
         }
     }
 
-    fn uptime_nano() -> u64 {
+    fn uptime_nano() -> u128 {
         let mut tp = libc::timespec {
             tv_sec: 0,
             tv_nsec: 0,
@@ -52,14 +52,14 @@ impl Delay for super::StdCommon {
         unsafe { libc::clock_gettime(libc::CLOCK_MONOTONIC, &mut tp) };
 
         let ptr = &raw mut TIME_START;
-        let t0 = unsafe { read_volatile(&*addr_of!((*ptr).tv_sec)) } * 1_000_000_000
-            + unsafe { read_volatile(&(*ptr).tv_nsec) };
-        let t1 = tp.tv_sec * 1_000_000_000 + tp.tv_nsec;
+        let t0 = unsafe { read_volatile(&*addr_of!((*ptr).tv_sec)) as u128 } * 1_000_000_000
+            + unsafe { read_volatile(&(*ptr).tv_nsec) as u128 };
+        let t1 = tp.tv_sec as u128 * 1_000_000_000 + tp.tv_nsec as u128;
 
         if t0 == 0 {
             0
         } else {
-            (t1 - t0) as u64
+            t1 - t0
         }
     }
